+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed ExtendedTreeView to ChapterTreeView
 - Renamed FormatterTextBlock to ChapterTextBlock
 - Renamed HeaderItemsControl to ChapterHeaderedContentControl
+- Renamed ImageButton to ChapterButton
 ### Supported .Net Versions
 - .Net Core 3.0
 - .Net Framework 4.5